The UserAssist information csv file is not generated
In line 66 of the method 'csv_user_assist_value_decode_win7_and_after'
in reg.py, the string argument has a length of 8 instead of 4 which is
the expected length if we consider the format argument provided to the
method 'unpack' ('<I').
To obtain the timestamp of the last exec date, the format argument
should be '<Q' to take into account the 8 bytes of the provided string.

Same issue in the method 'csv_user_assist_value_decode_before_win7' line
96